BurgerMenu Refactors
