+++ src/components/Layout/BurgerMenu/MobileNavigation.tsx
@@ -7,10 +7,10 @@ import prevent from 'decentraland-gatsby/dist/utils/react/prevent'
 import { Button } from 'decentraland-ui/dist/components/Button/Button'
 import { Header } from 'decentraland-ui/dist/components/Header/Header'
 
-import locations from '../../modules/locations'
+import locations from '../../../modules/locations'
+import { NavigationProps, NavigationTab } from '../Navigation'
 
 import './MobileNavigation.css'
-import { NavigationProps, NavigationTab } from './Navigation'
 
 function getProps(tab: NavigationTab, activeTab?: NavigationTab) {
   if (activeTab === tab) {
