+++ src/components/Layout/Navbar.tsx
@@ -4,7 +4,7 @@ import UserMenu from 'decentraland-gatsby/dist/components/User/UserMenu'
 import useResponsive from 'decentraland-gatsby/dist/hooks/useResponsive'
 import Responsive from 'semantic-ui-react/dist/commonjs/addons/Responsive'
 
-import BurgerMenu from './BurgerMenu'
+import BurgerMenu from './BurgerMenu/BurgerMenu'
 
 function Navbar({ props }: any) {
   const responsive = useResponsive()
