+++ src/pages/index.tsx
@@ -22,7 +22,8 @@ import RandomBanner from '../components/Banner/RandomBanner'
 import CategoryBanner from '../components/Category/CategoryBanner'
 import Empty from '../components/Common/Empty'
 import ActionableLayout from '../components/Layout/ActionableLayout'
-import BurgerMenuContent from '../components/Layout/BurgerMenuContent'
+import BurgerMenuContent from '../components/Layout/BurgerMenu/BurgerMenuContent'
+import BurgerMenuPushableLayout from '../components/Layout/BurgerMenu/BurgerMenuPushableLayout'
 import LoadingView from '../components/Layout/LoadingView'
 import Navigation, { NavigationTab } from '../components/Layout/Navigation'
 import ProposalItem from '../components/Proposal/ProposalItem'
@@ -65,7 +66,7 @@ export default function IndexPage() {
   const responsive = useResponsive()
   const isMobile = responsive({ maxWidth: Responsive.onlyMobile.maxWidth })
   const { status: burgerStatus } = useBurgerMenu()
-  const { open, translate } = burgerStatus
+  const { open } = burgerStatus
 
   const handlePageFilter = useCallback(
     (page: number) => {
@@ -163,72 +164,72 @@ export default function IndexPage() {
                   </div>
                 )}
               </Grid.Column>
-              <Grid.Column
-                tablet="12"
-                className="Animated ProposalsTable"
-                style={isMobile ? (translate ? { transform: `translateY(${translate})` } : {}) : {}}
-              >
-                {isMobile && proposals && <SearchTitle />}
-                <ActionableLayout
-                  leftAction={
-                    <Header sub>
-                      {!proposals && ''}
-                      {proposals && t('general.count_proposals', { count: proposals.total || 0 })}
-                    </Header>
-                  }
-                  rightAction={
-                    !searching && (
-                      <>
-                        {proposals && <SortingMenu />}
-                        <Button
-                          primary
-                          size="small"
-                          className="SubmitButton"
-                          as={Link}
-                          href={locations.submit()}
-                          onClick={prevent(() => navigate(locations.submit()))}
-                        >
-                          {t('page.proposal_list.new_proposal')}
-                        </Button>
-                      </>
-                    )
-                  }
-                >
-                  <Loader active={!proposals || isLoadingProposals} />
-                  {type && !searching && <CategoryBanner type={type} active />}
-                  {proposals && proposals.data.length === 0 && (
-                    <Empty
-                      description={
-                        searching || status || timeFrame?.length > 0
-                          ? t('navigation.search.no_matches')
-                          : t('page.proposal_list.no_proposals_yet')
-                      }
-                    />
-                  )}
-                  {proposals &&
-                    proposals.data.map((proposal) => {
-                      return (
-                        <ProposalItem
-                          key={proposal.id}
-                          proposal={proposal}
-                          votes={votes ? votes[proposal.id] : undefined}
-                          subscribing={subscriptionsState.subscribing.includes(proposal.id)}
-                          subscribed={!!subscriptions.find((subscription) => subscription.proposal_id === proposal.id)}
-                          onSubscribe={(_, proposal) => subscriptionsState.subscribe(proposal.id)}
-                        />
+              <BurgerMenuPushableLayout>
+                <Grid.Column tablet="12" className="ProposalsTable">
+                  {isMobile && proposals && <SearchTitle />}
+                  <ActionableLayout
+                    leftAction={
+                      <Header sub>
+                        {!proposals && ''}
+                        {proposals && t('general.count_proposals', { count: proposals.total || 0 })}
+                      </Header>
+                    }
+                    rightAction={
+                      !searching && (
+                        <>
+                          {proposals && <SortingMenu />}
+                          <Button
+                            primary
+                            size="small"
+                            className="SubmitButton"
+                            as={Link}
+                            href={locations.submit()}
+                            onClick={prevent(() => navigate(locations.submit()))}
+                          >
+                            {t('page.proposal_list.new_proposal')}
+                          </Button>
+                        </>
                       )
-                    })}
-                  {proposals && proposals.total > ITEMS_PER_PAGE && (
-                    <Pagination
-                      onPageChange={(e, { activePage }) => handlePageFilter(activePage as number)}
-                      totalPages={Math.ceil(proposals.total / ITEMS_PER_PAGE)}
-                      activePage={page}
-                      firstItem={null}
-                      lastItem={null}
-                    />
-                  )}
-                </ActionableLayout>
-              </Grid.Column>
+                    }
+                  >
+                    <Loader active={!proposals || isLoadingProposals} />
+                    {type && !searching && <CategoryBanner type={type} active />}
+                    {proposals && proposals.data.length === 0 && (
+                      <Empty
+                        description={
+                          searching || status || timeFrame?.length > 0
+                            ? t('navigation.search.no_matches')
+                            : t('page.proposal_list.no_proposals_yet')
+                        }
+                      />
+                    )}
+                    {proposals &&
+                      proposals.data.map((proposal) => {
+                        return (
+                          <ProposalItem
+                            key={proposal.id}
+                            proposal={proposal}
+                            votes={votes ? votes[proposal.id] : undefined}
+                            subscribing={subscriptionsState.subscribing.includes(proposal.id)}
+                            subscribed={
+                              !!subscriptions.find((subscription) => subscription.proposal_id === proposal.id)
+                            }
+                            onSubscribe={(_, proposal) => subscriptionsState.subscribe(proposal.id)}
+                          />
+                        )
+                      })}
+                    {proposals && proposals.total > ITEMS_PER_PAGE && (
+                      <Pagination
+                        onPageChange={(e, { activePage }) => handlePageFilter(activePage as number)}
+                        totalPages={Math.ceil(proposals.total / ITEMS_PER_PAGE)}
+                        activePage={page}
+                        firstItem={null}
+                        lastItem={null}
+                      />
+                    )}
+                  </ActionableLayout>
+                </Grid.Column>
+              </BurgerMenuPushableLayout>
             </Grid.Row>
           </Grid>
         </Container>
